fix(fp8_kernel): cast g_id to int64 to prevent overflow for large tensors
tl.program_id() returns int32. Multiplying by y_stride before adding to a
pointer can overflow for large tensors (long sequences or large hidden dims).
Cast to int64 before pointer arithmetic, consistent with the existing pattern
in _per_token_group_quant_8bit_colmajor at line 215.

+++ python/sglang/srt/layers/quantization/fp8_kernel.py
@@ -166,7 +166,7 @@ def _per_token_group_quant_8bit(
 
     This function converts the tensor values into float8 values.
     """
-    g_id = tl.program_id(0)
+    g_id = tl.program_id(0).to(tl.int64)
     y_off = g_id * y_stride
     y_ptr += y_off
     y_q_ptr += y_off
